Add Tabs in searchvue - begin
