+++ src/composables/components/search/SearchDocumentsForm.ts
@@ -1,10 +1,17 @@
-import { useElasticsearchAdapter } from '../../CallElasticsearch'
-import { useSearchStore } from '../../../store/search'
-import { useResizeStore } from '../../../store/resize'
-import { Ref, ref, watch } from 'vue'
-import { parseJson } from '../../../helpers/json/parse'
-import { DEFAULT_SEARCH_QUERY_OBJ } from '../../../consts'
+import { useElasticsearchAdapter } from '../../CallElasticsearch.ts'
+import { useSearchStore } from '../../../store/search.ts'
+import { useResizeStore } from '../../../store/resize.ts'
+import { nextTick, Ref, ref, toRaw, watch } from 'vue'
+import { parseJson } from '../../../helpers/json/parse.ts'
+import { DEFAULT_SEARCH_QUERY_OBJ } from '../../../consts.ts'
 import { stringifyJson } from '../../../helpers/json/stringify.ts'
+import { IdbSearchDocumentTab } from '../../../db/types.ts'
+import { debounce } from 'quasar'
+import { useIdbStore } from '../../../db/Idb.ts'
+
+type SearchDocumentormProps = {
+  tab: IdbSearchDocumentTab
+}
 
 export type EsSearchResult = {
   took: number | null
@@ -20,11 +27,20 @@ type EsSearchResultsHitsValues = {
   value: number
 }
 
-export const useSearchDocuments = () => {
+export const useSearchDocumentsForm = (props:SearchDocumentormProps) => {
+
+  const { searchDocumentTabs  } = useIdbStore()
+
+  const ownTab = ref(props.tab)
+  let updateIdb = true
+
   const { requestState, callElasticsearch } = useElasticsearchAdapter()
 
   const searchStore = useSearchStore()
   const resizeStore = useResizeStore()
+
+  searchStore.searchQuery = ownTab.value.query
+  searchStore.searchQueryCollapsed = ownTab.value.searchQueryCollapsed
 
   const searchResults: Ref<EsSearchResult> = ref({ took: null, hits: { total: { value: 0 }, hits: [] } })
   const queryParsingError = ref(false)
@@ -84,8 +100,44 @@ export const useSearchDocuments = () => {
 
   const mergeQuery = (params: any) => {
     const json = Object.assign({}, DEFAULT_SEARCH_QUERY_OBJ, params)
-    searchStore.searchQuery = stringifyJson(json, null, '\t')
+    searchStore.searchQuery = stringifyJson(json, null, '	')
   }
+
+  watch(() => searchStore.searchQuery, value => {
+    if (updateIdb) updateTab({ query: value })
+  })
+
+  watch(() => searchStore.searchQueryCollapsed, value => {
+    if (updateIdb) updateTab({ searchQueryCollapsed: value })
+  })
+
+  watch(ownTab.value.request, value => {
+    if (updateIdb) updateTab({ request: toRaw(value) })
+  })
+  watch(ownTab.value.response, value => {
+    if (updateIdb) updateTab({ response: toRaw(value) })
+  })
+  const updateTab = debounce((value: object) => {
+    const obj = Object.assign({}, toRaw(props.tab), value)
+    searchDocumentTabs.update(obj)
+  }, 100)
+
+  watch(() => props.tab, newValue => {
+    updateIdb = false
+    ownTab.value.query = newValue.query
+    ownTab.value.searchQueryCollapsed = newValue.searchQueryCollapsed
+    ownTab.value.request.method = newValue.request.method
+    ownTab.value.request.path = newValue.request.path
+    ownTab.value.request.body = newValue.request.body
+
+    searchStore.searchQuery = newValue.query
+    searchStore.searchQueryCollapsed = newValue.searchQueryCollapsed
+
+    nextTick(() => {
+      updateIdb = true
+    })
+  })
+    
 
   const editorCommands = [
     {
