+++ src/composables/components/search/SearchResultsTable.ts
@@ -5,7 +5,7 @@ import { useElasticsearchAdapter } from '../../CallElasticsearch.ts'
 import { useSelectableRows } from '../../SelectableRow.ts'
 import SearchResults from '../../../models/SearchResults.ts'
 import { sortableField } from '../../../helpers/search.ts'
-import { EsSearchResult } from './SearchDocuments.ts'
+import { EsSearchResult } from './SearchDocumentsForm.ts'
 import { ElasticsearchDocumentInfo } from './EditDocument.ts'
 import { filterItems } from '../../../helpers/filters.ts'
 import { stringifyJson } from '../../../helpers/json/stringify.ts'
